ci: publish project to PyPI in the same job that we release (#68)
This allows the publish step to re-use the permissions
given to the release step, namely the admin's authorization
to run jobs in the `release` environment.

## Summary by Sourcery

Consolidates the release and PyPI publishing jobs into a single job to
reuse permissions, and updates the uv-pre-commit hook version.

CI:
- Combines the release and PyPI publishing jobs into a single job within
the CI workflow to streamline the release process.
- Updates the uv-pre-commit hook version to 0.6.6 in
.pre-commit-config.yaml file.